Add files via upload
We've also prepared an ai model which we're looking to incorporate in our app. The model is a cnn based binary classifier which predicts the input image to be benign or malignant. Due to the low availability of data our model might not be perfect rn but we're wat we're looking at is a lot of scope for improvement.